+++ include/flags/flagsfwd.hpp
@@ -1,0 +1,8 @@
+#ifndef ENUM_CLASS_FLAGSFWD_HPP
+#define ENUM_CLASS_FLAGSFWD_HPP
+
+
+namespace flags { template <class E> struct flags; }
+
+
+#endif // ENUM_CLASS_FLAGSFWD_HPP
